Fix: Update uuid_id type from i16 to i32 across multiple modules for improved compatibility

+++ server/src/entity/dynamic_monitoring.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i64,
-    pub uuid_id: i16,
+    pub uuid_id: i32,
     pub timestamp: i64,
     #[sea_orm(column_type = "JsonBinary")]
     pub cpu_data: Json,
