+++ server/src/rpc/agent/query_dynamic_summary.rs
@@ -121,7 +121,7 @@ pub async fn query_dynamic_summary(
         let mut is_last = false;
 
         // Pre-resolve UUID conditions to uuid_id via cache
-        let mut uuid_ids: Vec<i16> = Vec::new();
+        let mut uuid_ids: Vec<i32> = Vec::new();
         for cond in &query_data.condition {
             if let QueryCondition::Uuid(uuid) = cond {
                 let uuid_id = uuid_cache.get_id(uuid).await.ok_or_else(|| {
@@ -285,11 +285,11 @@ async fn execute_query(
             Ok(mut v) => {
                 result_count += 1;
 
-                // Translate uuid_id (i16) → uuid (string) for API compatibility
+                // Translate uuid_id (i32) → uuid (string) for API compatibility
                 if let Value::Object(ref mut map) = v {
                     if let Some(Value::Number(n)) = map.remove("uuid_id") {
                         if let Some(id) = n.as_i64() {
-                            if let Some(uuid) = uuid_cache.get_uuid(id as i16).await {
+                            if let Some(uuid) = uuid_cache.get_uuid(id as i32).await {
                                 map.insert("uuid".to_string(), Value::String(uuid.to_string()));
                             }
                         }
