+++ server/src/rpc/agent/query_dynamic_summary_multi_last.rs
@@ -93,7 +93,7 @@ pub async fn dynamic_summary_multi_last_query(
         let uuid_cache = MonitoringUuidCache::global();
 
         // Resolve UUIDs to uuid_ids
-        let mut uuid_id_pairs: Vec<(Uuid, i16)> = Vec::with_capacity(deduped_uuids.len());
+        let mut uuid_id_pairs: Vec<(Uuid, i32)> = Vec::with_capacity(deduped_uuids.len());
         for uuid in &deduped_uuids {
             let uuid_id = uuid_cache.get_id(uuid).await.ok_or_else(|| {
                 NodegetError::NotFound(format!(
@@ -142,7 +142,7 @@ fn dedupe_uuids(uuids: Vec<Uuid>) -> Vec<Uuid> {
 }
 
 fn build_union_last_statement(
-    uuid_id_pairs: &[(Uuid, i16)],
+    uuid_id_pairs: &[(Uuid, i32)],
     fields: &[DynamicSummaryQueryField],
     db: &DatabaseConnection,
 ) -> anyhow::Result<Statement> {
@@ -169,7 +169,7 @@ fn is_scaled_column(name: &str) -> bool {
     SCALED_COLUMNS.contains(&name)
 }
 
-fn build_single_last_select(uuid_id: i16, fields: &[DynamicSummaryQueryField]) -> SelectStatement {
+fn build_single_last_select(uuid_id: i32, fields: &[DynamicSummaryQueryField]) -> SelectStatement {
     let inner_query = dynamic_monitoring_summary::Entity::find()
         .select_only()
         .column(dynamic_monitoring_summary::Column::UuidId)
@@ -270,7 +270,7 @@ async fn execute_statement_query(
                 if let Some(obj) = value.as_object_mut() {
                     if let Some(uuid_id_val) = obj.remove("uuid_id") {
                         if let Some(uuid_id) = uuid_id_val.as_i64() {
-                            if let Some(uuid) = uuid_cache.get_uuid(uuid_id as i16).await {
+                            if let Some(uuid) = uuid_cache.get_uuid(uuid_id as i32).await {
                                 obj.insert(
                                     "uuid".to_owned(),
                                     serde_json::Value::String(uuid.to_string()),
